fix for windows
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@2786 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ sip/resiprocate/os/Data.cxx
@@ -711,7 +711,7 @@ Data
 Data::operator+(const char* str) const
 {
    assert(str);
-   unsigned int l = strlen(str);
+   size_t l = strlen(str);
    Data tmp(mSize + l, true);
    tmp.mSize = mSize + l;
    tmp.mCapacity = tmp.mSize;
